+++ drizzle/0002_luxuriant_maginty.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `video_projects` ADD `is_demo` integer DEFAULT false NOT NULL;